@@ -35,7 +35,7 @@
     "cook-eslint": "index.js"
   },
   "peerDependencies": {
-    "eslint": "^5.0.0 || ^6.0.0 || ^7.0.0"
+    "eslint": "^7.0.0"
   },
   "dependencies": {
     "commander": "^8.2.0",
